requirements.txt
amqp==5.2.0
APScheduler==3.10.4
asgiref==3.8.1
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
billiard==4.2.0
celery==5.4.0
certifi==2024.6.2
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cron-descriptor==1.4.3
cryptography==42.0.8
daphne==4.1.2
distlib==0.3.8
Django==5.0.6
django-apscheduler==0.6.2
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-cors-headers==4.4.0
django-environ==0.11.2
django-redis==5.4.0
django-timezone-field==7.0
djangorestframework==3.15.2
drf-yasg==1.21.7
filelock==3.15.4
gTTS==2.5.1
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
inflection==0.5.1
jsonfield==3.1.0
kombu==5.3.7
msgpack==1.0.8
mysqlclient==2.2.4
packaging==24.1
platformdirs==4.2.2
prompt_toolkit==3.0.47
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pyOpenSSL==24.1.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.1
redis==5.0.7
requests==2.32.3
service-identity==24.1.0
six==1.16.0
sqlparse==0.5.0
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.2
views==0.3
vine==5.1.0
virtualenv==20.26.3
wcwidth==0.2.13
zope.interface==6.4.post2
